Added Title bar to control
In user testing, it was apparent that users were initially unaware of
the pairing of the markers they clicked on and the change of the
information modules in the control panel. By adding a changing title to
the top of the control bar, we hope that users will be more aware of
their relationship.